refactor: split api.ts into separate modules and add base option validation
- Split packages/github/src/api.ts into separate modules under api/ directory
  - api/types.ts: interfaces and type guard
  - api/repo-info.ts: getGitHubRepoInfo function
  - api/pull-request.ts: fetchPullRequest function
  - api/issue.ts: fetchIssue function
  - api/index.ts: re-exports all functions and types

- Add validation to prevent --base option with pull requests
  - Modified githubCheckout to return Result<void>
  - Returns error when base is specified for PRs with clear message
  - Updated CLI handler to use exitWithError for proper error handling
