+++ packages/github/src/api/index.ts
@@ -1,0 +1,4 @@
+export * from "./types.ts";
+export * from "./repo-info.ts";
+export * from "./pull-request.ts";
+export * from "./issue.ts";
