+++ packages/github/src/api/pull-request.ts
@@ -1,0 +1,32 @@
+import { z } from "zod";
+import { createGitHubClient } from "../client.ts";
+import type { GitHubPullRequest } from "./types.ts";
+
+const numberSchema = z.coerce.number().int().positive();
+
+export async function fetchPullRequest(
+  owner: string,
+  repo: string,
+  number: string,
+): Promise<GitHubPullRequest | null> {
+  try {
+    const pullNumber = numberSchema.parse(number);
+    const octokit = await createGitHubClient();
+    const { data } = await octokit.pulls.get({
+      owner,
+      repo,
+      pull_number: pullNumber,
+    });
+    return {
+      number: data.number,
+      head: {
+        ref: data.head.ref,
+      },
+    };
+  } catch (error) {
+    if (error instanceof Error && "status" in error && error.status === 404) {
+      return null;
+    }
+    throw error;
+  }
+}
